Refactor code and Implement coments
