+++ src/main/java/co/edu/uniquindio/unieventos/dto/DTOCrearCuenta.java
@@ -1,0 +1,19 @@
+package co.edu.uniquindio.unieventos.dto;
+
+public record DTOCrearCuenta(
+
+        // TODO Implementar datos del DTO
+
+    //TODO este es un ejemplo de creacion de DTO se tiene que hacer sobre nuestro modelo
+
+        /*
+        @NotBlank @Length(max = 10) String cedula,
+        @NotBlank @Length(max = 100) String nombre,
+        @NotBlank @Length(max = 10) String telefono,
+        @Length(max = 100) String direccion,
+        @NotBlank @Length(max = 50) @Email String email,
+        @NotBlank @Length(min = 7, max = 20) String password*/
+) {
+
+
+}
