--- src/main/java/co/edu/uniquindio/unieventos/modelo/vo/HistorialActividad.java
@@ -1,20 +1,0 @@
-package co.edu.uniquindio.unieventos.modelo.vo;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-import java.util.Date;
-
-
-@Getter
-@Setter
-@NoArgsConstructor
-@ToString
-public class HistorialActividad {
-
-    private Date fechaActividad;
-    private String descripcion;
-
-}
