+++ src/main/java/co/edu/uniquindio/unieventos/repositorio/CarritoComprasRepository.java
@@ -6,5 +6,4 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CarritoComprasRepository extends MongoRepository<CarritoCompras,String> {
-
 }
