+++ src/main/java/co/edu/uniquindio/unieventos/servicios/interfases/CuentaServicio.java
@@ -1,9 +1,9 @@
 package co.edu.uniquindio.unieventos.servicios.interfases;
 
-import co.edu.uniquindio.unieventos.modelo.documentos.Cuenta;
+import co.edu.uniquindio.unieventos.dto.DTOCrearCuenta;
 
 public interface CuentaServicio {
 
-    default void crearCuenta(DtoCuenta cuenta){}
+    String crearCuenta(DTOCrearCuenta cuenta) throws Exception;
 
 }
