--- src/test/java/co/edu/uniquindio/proyecto/test/CarritoComprasTest.java
@@ -1,4 +1,0 @@
-package co.edu.uniquindio.proyecto.test;
-
-public class CarritoComprasTest {
-}
