--- src/test/java/co/edu/uniquindio/proyecto/test/CompraTest.java
@@ -1,4 +1,0 @@
-package co.edu.uniquindio.proyecto.test;
-
-public class CompraTest {
-}
